+++ packages/devtools_app/lib/src/shared/file_import.dart
@@ -5,7 +5,6 @@
 import 'dart:convert';
 
 import 'package:devtools_app_shared/ui.dart';
-import 'package:devtools_app_shared/utils.dart';
 import 'package:file_selector/file_selector.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
